View Grade
Fix the display of the grade table of students .
Added display of average grade and subject status.
Test cases of average grade:
1. The total grade of the grade components is not enough -> Studying
a. There is no total equal to 0 -> Studying
b. There is a total of 0 -> Studying
2. Have enough total grades of the components and No retake exam, no final exam resist grade
a. All component grades are enough to pass
b. The average grade is above 5 but has a total of 0
c. The average grade is above 5 but the final exam total is below 4
d. The average grade is below 5 but the final exam total is above 4
e. Average grade below 5 and total final exam grade below 4
3. Have enough total grades of the score components and have exam retake, no final exam resist grade
a. All component grades are enough to pass (final exam above 4), regardless of final exam grade.
b. All component grades are enough to pass (final exam grade below 4), regardless of final exam grade.
c. The average grade is above 5 but has a total of 0
d.The average grade is above 5 but the final exam total is below 4
e. The average grade is below 5 but the final exam total is above 4
f. Average grade below 5 and total final exam grade below 4